requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
duckdb>=1.0.0
psycopg2-binary>=2.9.9

# ONNX/AI
onnxruntime>=1.18.0
tokenizers>=0.19.0
numpy>=1.26.0

# Utilities
httpx>=0.27.0
tqdm>=4.66.0
rich>=13.7.0
click>=8.1.0
fsrs>=4.1.0

# Dev
pytest>=8.0.0
pytest-asyncio>=0.23.0
